backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Flotschi <tanicruzzer2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
django-allauth = "^0.54.0"
pyyaml = "^6.0.2"
drf-yasg = "^1.21.8"
requests = "^2.32.3"
pytest-django = "^4.9.0"
mixer = "^7.2.2"
dj-rest-auth = "^7.0.0"
coverage = "^7.6.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
